+++ gdsfactory/geometry/write_drc.py
@@ -1,4 +1,10 @@
-"""Write DRC rule decks in klayout."""
+"""Write DRC rule decks in klayout.
+
+TODO:
+
+- add min area
+- define derived layers (composed rules)
+"""
 
 import pathlib
 from dataclasses import asdict, is_dataclass
@@ -25,7 +31,7 @@ RuleType = Literal[
 def rule_width(value: float, layer: str, angle_limit: float = 90) -> str:
     """Min feature size"""
     category = "width"
-    error = f"{layer} {category} {value}"
+    error = f"{layer} {category} {value}um"
     return (
         f"{layer}.{category}({value}, angle_limit({angle_limit}))"
         f".output('{error}', '{error}')"
@@ -35,7 +41,7 @@ def rule_width(value: float, layer: str, angle_limit: float = 90) -> str:
 def rule_space(value: float, layer: str, angle_limit: float = 90) -> str:
     """Min Space between shapes of layer"""
     category = "space"
-    error = f"{layer} {category} {value}"
+    error = f"{layer} {category} {value}um"
     return (
         f"{layer}.{category}({value}, angle_limit({angle_limit}))"
         f".output('{error}', '{error}')"
@@ -44,19 +50,20 @@ def rule_space(value: float, layer: str, angle_limit: float = 90) -> str:
 
 def rule_separation(value: float, layer1: str, layer2: str):
     """Min space between different layers"""
-    error = f"min {layer1} {layer2} separation {value}"
+    error = f"min {layer1} {layer2} separation {value}um"
     return f"{layer1}.separation({layer2}, {value})" f".output('{error}', '{error}')"
 
 
 def rule_enclosing(
     value: float, layer1: str, layer2: str, angle_limit: float = 90
 ) -> str:
-    """Layer1 must be enclosed by layer2 by value"""
-    category = "enclosing"
-    error = f"{layer1} {category} {value}"
+    """Layer1 must be enclosed by layer2 by value.
+    checks if layer1 encloses (is bigger than) layer2 by value
+    """
+    error = f"{layer1} enclosing {layer2} by {value}um"
     return (
-        f"{layer1}.{category}({layer2}, angle_limit({angle_limit}), {value})"
-        f".output('{error}', '{layer2} minimum {category} {value}')"
+        f"{layer1}.enclosing({layer2}, angle_limit({angle_limit}), {value})"
+        f".output('{error}', '{error}')"
     )
 
 
@@ -89,12 +96,20 @@ def write_drc_deck(rules: List[str], layer_map: Dict[str, Layer]) -> str:
 
 
 def write_drc_deck_macro(
-    name="generic", filepath: Optional[PathType] = None, **kwargs
+    name="generic",
+    filepath: Optional[PathType] = None,
+    shortcut: str = "Ctrl+Shift+D",
+    **kwargs,
 ) -> str:
     """Write script for klayout rule deck
 
     Args:
         name: drc rule deck name
+        filepath: Optional macro path (defaults to .klayout/drc/name.lydrc)
+
+    Keyword Args:
+        rules: list of rules
+        layer_map: layer definitions can be dict or dataclass
 
     Keyword Args:
         rules: list of rules
@@ -111,7 +126,7 @@ def write_drc_deck_macro(
  <doc/>
  <autorun>false</autorun>
  <autorun-early>false</autorun-early>
- <shortcut>Ctrl+Shift+D</shortcut>
+ <shortcut>{shortcut}</shortcut>
  <show-in-menu>true</show-in-menu>
  <group-name>drc_scripts</group-name>
  <menu-path>tools_menu.drc.end</menu-path>
@@ -146,11 +161,12 @@ if __name__ == "__main__":
 
     rules = [
         rule_width(layer="WG", value=0.2),
+        rule_space(layer="WG", value=0.2),
         rule_width(layer="M1", value=1),
         rule_width(layer="M2", value=2),
-        rule_space(layer="WG", value=0.2),
         rule_space(layer="M2", value=2),
-        rule_enclosing(layer1="M2", layer2="M1", value=2),
+        rule_separation(layer1="HEATER", layer2="M1", value=1.0),
+        rule_enclosing(layer1="M1", layer2="VIAC", value=0.2),
     ]
 
     drc_rule_deck = write_drc_deck_macro(rules=rules, layer_map=gf.LAYER)